* release/0.2:
  Updated CONTRIBUTORS for at-launch non-monetary contributors and added AUTHORS, which could probably be scripted based on the git commit history.
  Updated license header to correctly attribute copyright to Readium Foundation and contributors.
  there was a merge problem, so trying to sort out the submodule state...
  Added DAISY Consortium to list of contributors. NOTE: working from develop branch of SDK submodule from LauncherOSX.
  Updated copyright and contributor information.
  Removed curl for iOS (no longer used) and tweaked build settings.
  Moved Xcode project files into Platform/Apple. New platform-specific build files will be placed inside this folder.
  Implemented some basic localization features, primarily aimed at handling alternate-script metadata extensions.
  Fixed problem where iOS Release builds were defining USING_ICU == 1
  Fixed iOS linker error. • Had forgotten I was in a C++ file when declaring the external ObjC function. Added "C" to the extern declaration so all should be working now.
  Another error from the static analyzer that I thought I'd fixed in the last round.
  Fixed a couple of errors flagged by the static analyzer.
  Solved issues causing build failure for iOS devices (including implicit cast warnings).
  Add ring_buffer.cpp to the iOS target.
  A few API-breaking changes: • Locator is gone. Use paths to access ePub archives and epub3 (i.e. IRI::gEPUBScheme) URLs to access documents. • New ByteStream API has covers in the most useful places. The following all return an appropriate synchronous ByteStream instance:   • ManifestItem::Reader() [now returns a ByteStream instead of an ArchiveReader].   • Library::ReadStreamForEPubURL() [new API].   • PackageBase::ReadStreamForItemAtPath() [new API].   • Package::ReadStreamForRelativePath() [new API].   • Container::ReadStreamAtPath() [new API].   • Archive::ByteStreamAtPath() [new API, implemented by ZipArchive]. • You can ask a Library object to give you a ByteStream for a full epub3:// URL, provided the requisite package has been loaded into the MainLibrary. This includes URLs pointing directly at sub-paths of the container archive and URLs containing only a CFI, i.e. epub3://<unique-id>/#epubcfi(/6/4[ch01]!). Support for file:// URLs directed at .epub files with CFIs is not yet supported, but is on the list.

Conflicts:
	CONTRIBUTORS

+++ ePub3/ePub/cfi.cpp
@@ -3,7 +3,7 @@
 //  ePub3
 //
 //  Created by Jim Dovey on 2012-12-10.
-//  Copyright (c) 2012-2013 The Readium Foundation.
+//  Copyright (c) 2012-2013 The Readium Foundation and contributors.
 //  
 //  The Readium SDK is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU General Public License as published by
@@ -382,7 +382,7 @@ void CFI::Component::Parse(const string &str)
         {
             case '[':
             {
-                size_t pos = iss.tellg();
+                size_t pos = static_cast<size_t>(iss.tellg());
                 iss.ignore(std::numeric_limits<std::streamsize>::max(), ']');
                 size_t end = ((size_t)iss.tellg()) - 1;
